refactor(metrics): move provider implementations into providers/ subfolder
Move meter-provider.ts, prometheus-meter-provider.ts, and
prometheus-metrics-factory.ts into src/providers/ to align with the
standard package folder structure documented in
docs/reference/package-folder-structure.md.

Update all internal import paths within the package. No business logic,
type, or interface changes. index.ts public API re-export signatures
remain identical.

Closes #859
