+++ packages/metrics/src/index.ts
@@ -1,6 +1,6 @@
 export { Registry } from 'prom-client';
 export * from './metrics-module.js';
 export * from './metrics-service.js';
-export * from './meter-provider.js';
-export * from './prometheus-meter-provider.js';
+export * from './providers/meter-provider.js';
+export * from './providers/prometheus-meter-provider.js';
 export * from './http-metrics-middleware.js';
